Crate server_fn_macro

Implementation of the server_fn macro.

This crate contains the implementation of the server_fn macro. server_macro_impl can be used to implement custom versions of the macro for different frameworks that allow users to pass a custom context from the server to the server function.

Structs

ServerContext

Describes the custom context from the server that passed to the server function. Optionally, the first argument of a server function can be a custom context of this type. This context can be used to access the server’s state within the server function.

Functions

server_macro_impl

The implementation of the server_fn macro. To allow the macro to accept a custom context from the server, pass a custom server context to this function. The Context comes from the server. Optionally, the first argument of a server function can be a custom context. This context can be used to inject dependencies like the HTTP request or response or other server-only dependencies, but it does not have access to state that exists in the client.